refactor(api): move competence-evaluation.js model
