+++ api/lib/domain/services/scorecard-service.js
@@ -1,5 +1,5 @@
 import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { CompetenceEvaluation } from '../models/CompetenceEvaluation.js';
+import { CompetenceEvaluation } from '../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 import { KnowledgeElement } from '../models/KnowledgeElement.js';
 import { Scorecard } from '../models/Scorecard.js';
 import _ from 'lodash';
