+++ api/lib/infrastructure/repositories/competence-evaluation-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { NotFoundError } from '../../domain/errors.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import { knex } from '../../../db/knex-database-connection.js';
-import { CompetenceEvaluation } from '../../domain/models/CompetenceEvaluation.js';
+import { CompetenceEvaluation } from '../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
 
 const save = async function ({ competenceEvaluation, domainTransaction = DomainTransaction.emptyTransaction() }) {
