+++ api/src/evaluation/application/competence-evaluations/competence-evaluation-controller.js
@@ -1,4 +1,4 @@
-import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { evaluationUsecases as usecases } from '../../../evaluation/domain/usecases/index.js';
 import * as competenceEvaluationSerializer from '../../../../lib/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
