+++ api/tests/evaluation/unit/domain/usecases/start-or-resume-competence-evaluation_test.js
@@ -1,8 +1,9 @@
 import { expect, sinon, catchErr } from '../../../../test-helper.js';
-import { Assessment, CompetenceEvaluation } from '../../../../../lib/domain/models/index.js';
 import { evaluationUsecases as usecases } from '../../../../../src/evaluation/domain/usecases/index.js';
 import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import _ from 'lodash';
+import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
+import { CompetenceEvaluation } from '../../../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 
 describe('Unit | UseCase | start-or-resume-competence-evaluation', function () {
   const userId = 123;
