+++ api/tests/integration/infrastructure/repositories/competence-evaluation-repository_test.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
-import { CompetenceEvaluation } from '../../../../lib/domain/models/CompetenceEvaluation.js';
+import { CompetenceEvaluation } from '../../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import * as competenceEvaluationRepository from '../../../../lib/infrastructure/repositories/competence-evaluation-repository.js';
 import { catchErr, databaseBuilder, expect, knex } from '../../../test-helper.js';
