+++ api/tests/tooling/domain-builder/factory/build-competence-evaluation.js
@@ -1,4 +1,4 @@
-import { CompetenceEvaluation } from '../../../../lib/domain/models/CompetenceEvaluation.js';
+import { CompetenceEvaluation } from '../../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 import { buildAssessment } from './build-assessment.js';
 import { buildUser } from './build-user.js';
